Change method of logging in.
As I learned more about Oauth, I realized that having a user sign into an account via a third party
application is incredibly insecure. I learned the proper methods of authenticating a user on my
application. As a result, right now, all the user needs to do is sign on via the Shapeways website,
and give me the pin they obtained.