Change logic for obtaining 3d files.
Rather than attempting to obtain an stl file from the shapeJS server and convert it to a g3db on the phone (which is taxing and a bad idea in the first place, not to mention extremely difficult), the application will communicate with a custom server. That custom server will retrieve an image file, communicate with shapeJS, convert the received 3d file (twice), and return it to the application. A link to the server code will be provided within the Readme, along with the utilized conversion libraries.