Integrate code into a LibGDX project.
I'm going to have to apologize for this massive disruption in project history. When I began this project, I thought it would be easier to simply program all the android specific code first, and then transitionit into a LibGDX project. I did not have clean project history in mind when I made this decision. This is still essentially the same application as the previous commit.

Fixes #25