Redesign model storage architecture.
I just realized that in order to submit the order on shapeways, I would need both the g3db file for
viewing, in addition to the stl file for uploading.

Now every model will have it's own folder with a unique name, typically the name of the image.
The folder will now contain both the g3db file as well as the stl file.