Fixes #34
Having a scrollview as the absolute parent to all fragments was pretty stupid.
It makes much more sense to add a scrollview when it's necessary.
Also having a listview inside a scrollview is widely discouraged among the android community, and
for good reason too. They both have their own scrolling events to take care of, and it'd be weird to
have them both react to the same scroll event.