Implement call to add to cart.
User must click the buy button once in order to upload it, wait a few minutes, and then click it again to actually add it to the cart.
This implementation is only temporary, and will soon be replaced with a cart page will allow you to select materials and checkout once the upload on that item is complete.
Fixes #27.