Fix bug on configuration changes.
When configurations were changed, e.g. Camera was used and the screen rotated, the reference to the adapter was lost and instant updates to the recyclerview were impossible. By making the adapter static, this issue is resolved.

Also changed the file directory to store models to the default external file storage.