Implement title editing
After two days of attempting to rename the files, for some strange reason, android refuses to
properly rename the files within the model directories. My workaround is simply storing the original
name the model was created with in addition to holding a display name, which will be kept in
SharedPreferences.