Implement Runtime Permission Library
Well the temporary solution of compiling to API 22 failed. Turns out you're not allowed to downgrade
from API 23 due to the fact that the permission logic was already enforced for Marshmallow users.
The Runtime Permissions really irked me and I despised the fact that it bloats the codebase. So I
used a library that autogenerates classes for me to handle all that crap. I hope Google considers
the developers point of views in future updates.